[TEST] Cover Gluon async copy simulation ops
